Fix pep8 violations.
Fixes:

* leaderboards/grade_source/__init__.py:57:1: E402 module level import not at top of file
* leaderboards/grade_source/__init__.py:58:1: E402 module level import not at top of file
* leaderboards/grade_source/edx.py:63:9: E731 do not assign a lambda expression, use a def
* leaderboards/grade_source/mock.py:44:9: E731 do not assign a lambda expression, use a def
* leaderboards/grade_source/mock.py:48:9: E731 do not assign a lambda expression, use a def
